move functions inline
